1_docker/15_python_script_azure_1/pyproject.toml
[project]
name = "15-python-script-azure-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-storage-blob>=12.25.1",
    "python-dotenv>=1.1.0",
]
